Updated Dialog and Education
- Updated Dialog to include esc key functionality
- Updated details on Education section
