+++ src/pages/Home.jsx
@@ -52,7 +52,7 @@ const Home = () => {
               <LocationIcon />
             </Icon>
             <div style={{ overflow: "hidden" }}>
-              <LocationText isHovered={isHovered}>
+              <LocationText style={{ left: isHovered ? "0" : "-100%" }}>
                 London,
                 <br />
                 United Kingdom
@@ -132,14 +132,15 @@ const Subtitle = styled(SecondaryTextStyle)`
 
 const LocationText = styled(SecondaryTextStyle)`
   position: relative;
-  left: ${({ isHovered }) => (isHovered ? "0" : "-100%")};
+  background-image: var(--grey-gradient);
+  color: transparent;
+  background-clip: text;
   transition: left 0.3s ease-in-out;
   margin: 0;
   font-size: 1rem;
   line-height: 1.1;
   overflow: hidden;
   letter-spacing: 2px;
-  color: var(--grey-dark-theme);
   @media (max-width: 768px) {
     font-size: 0.8rem;
   }
